[TM4C123_ACOMP] Updated casting in conditional Actualizar org.eclipse.core.resources.prefs, Buzzer_Init.c y 24 más archivos...

+++ TM4C123_ACOMP/xUtils/DataStructure/HashTable/ChainedHashTable/xSource/CHashTable_Insert.c
@@ -49,7 +49,7 @@ CHashTable_nSTATUS CHashTable__enInsert(CHashTable_TypeDef* psCHashTable, void* 
             psList = psCHashTable->psTable;
             psList += u32BucketNum;
             psElement = SLinkedList__psAddBegin(psList, pvData);
-            if((uint32_t)0UL != psElement)
+            if((uint32_t)0UL != (uint32_t)psElement)
             {
                 u32SizeReg = CHashTable__u32GetSize(psCHashTable);
                 u32SizeReg++;
