+++ TM4C123_ACOMP/xUtils/DataStructure/LinkedList/CircularDoubleLinkedList/xSource/CDLinkedList_Node.c
@@ -120,7 +120,7 @@
      uint32_t u32SizeReg = 0UL;
      uint32_t u32SizeList = 0UL;
 
-     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != pvData ) && (0UL != u32MaxLength))
+     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != (uint32_t)pvData ) && (0UL != u32MaxLength))
      {
          psMember = CDLinkedList__psGetHead(psList);
          u32SizeList = CDLinkedList__u32GetSize(psList);
@@ -146,7 +146,7 @@
      uint32_t u32SizeReg = 0UL;
      uint32_t u32SizeList = 0UL;
 
-     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != pvData ) && (0UL != u32MaxLength))
+     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != (uint32_t)pvData ) && (0UL != u32MaxLength))
      {
          psMember = CDLinkedList__psGetTail(psList);
          u32SizeList = CDLinkedList__u32GetSize(psList);
@@ -172,7 +172,7 @@
      uint32_t u32SizeReg = 0UL;
      uint32_t u32SizeList = 0UL;
 
-     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != pvData ) && (0UL != u32MaxLength)&& (0UL != u32Nodes))
+     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != (uint32_t)pvData ) && (0UL != u32MaxLength)&& (0UL != u32Nodes))
      {
          psMember = CDLinkedList__psGetHead(psList);
          u32SizeList = CDLinkedList__u32GetSize(psList);
@@ -199,7 +199,7 @@
      uint32_t u32SizeReg = 0UL;
      uint32_t u32SizeList = 0UL;
 
-     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != pvData ) && (0UL != u32MaxLength)&& (0UL != u32Nodes))
+     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL !=(uint32_t) pvData ) && (0UL != u32MaxLength)&& (0UL != u32Nodes))
      {
          psMember = CDLinkedList__psGetTail(psList);
          u32SizeList = CDLinkedList__u32GetSize(psList);
